Move gatsby-node code to plugin
