+++ plugins/gatsby-yummy-pages/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "gatsby-yummy-pages",
+  "version": "1.0.0"
+}
